Guard platforms better when not installed.

+++ boilerplate/scripts/helpers/buildWeb.js
@@ -1,18 +1,28 @@
+const gutil = require('gulp-util');
 const path = require('path');
 const rimraf = require('rimraf');
 const shell = require('shelljs');
 
+const tryPackage = require('./tryPackage');
+
 function buildWeb(callback) {
-  // Remove last build
-  rimraf(path.join(
-    process.cwd(),
-    'build',
-    'web',
-    global.settings.production ? 'production' : 'debug'
-  ), () => {
-    shell.exec(`npm run build:${global.settings.production ? 'production' : 'debug'}:web`);
+  const enabled = tryPackage('webpack');
+  if (enabled) {
+    // Remove last build
+    rimraf(path.join(
+      process.cwd(),
+      'build',
+      'web',
+      global.settings.production ? 'production' : 'debug'
+    ), () => {
+      shell.exec(`npm run build:${global.settings.production ? 'production' : 'debug'}:web`);
+      callback();
+    });
+  }
+  else {
+    gutil.log(gutil.colors.red('Need web platform. Run `gulp enable web`.'));
     callback();
-  });
+  }
 }
 
 module.exports = buildWeb;
